Changed the layerin of componenets so that Navigation is the root, followed by TabBar
